migrate test_base.py to test_raw.py
